move test files
add faker package for testing
change go-playground assert testing to stretchr testing
