+++ __test__/user_test.go
@@ -1,57 +1,78 @@
 package main
 
 import (
-	"api/controllers/users"
+	controller "api/controllers"
+	user "api/controllers/users"
+	"api/repository"
+	"api/service"
 	"bytes"
 	"encoding/json"
-	"log"
+	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"strconv"
 	"testing"
+
+	"github.com/go-faker/faker/v4"
+	"github.com/stretchr/testify/assert"
 )
 
-func tearDown(id string) {
+var (
+	userRepository = repository.NewUserRepository()
+	userService    = service.NewUserService(userRepository)
+	UserController = user.NewUserController(userService)
+	authController = controller.NewAuthController(userService)
+)
+
+func tearDown(t *testing.T, id string) {
 	var jsonStr = []byte(`{"ids": [` + id + `]}`)
 	request, _ := http.NewRequest("DELETE", "/v1/users/delete", bytes.NewBuffer(jsonStr))
 	request.Header.Set("Content-Type", "application/json")
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		users.Delete(w, r)
+		UserController.Delete(w, r)
 	})
 
-	executeRequest(request, handler)
+	response := executeRequest(request, handler)
+
+	var m map[string]any
+	err := json.Unmarshal(response.Body.Bytes(), &m)
+
+	if err != nil {
+		t.Errorf("Failed to delete user with ID" + id)
+	}
 }
 
 func TestLogin(t *testing.T) {
 
-	response := httptest.NewRecorder()
-	var jsonStr = []byte(`{"email":"rowel1@gmail.com", "passwords": "admins"}`)
-	request, err := http.NewRequest("POST", "/v1/auth/login", bytes.NewBuffer(jsonStr))
+	var jsonStr = []byte(`{"email":"rowel@gmail.com", "password": "admin"}`)
+	request, _ := http.NewRequest("POST", "/v1/auth/login", bytes.NewBuffer(jsonStr))
 	request.Header.Set("Content-Type", "application/json")
+	handler := http.HandlerFunc(authController.Login)
+
+	response := executeRequest(request, handler)
+
+	var m map[string]any
+	err := json.Unmarshal(response.Body.Bytes(), &m)
 
 	if err != nil {
-		t.Fatal("Creating 'GET /questions/1/SC' request failed!")
+		t.Errorf("Unexpected response body found")
 	}
-
-	app.Router.ServeHTTP(response, request)
-
-	if response.Code != http.StatusOK {
-		t.Fatal("Server error: Returned ", response.Code, " instead of ", http.StatusOK)
+	if m["statusCode"] != float64(200) {
+		t.Errorf("Wrong status code. expecting 200, receive %v", m["statusCode"])
 	}
 }
 
 func TestCreates(t *testing.T) {
 
-	var jsonStr = []byte(`{"firstName":"Rowel", "lastName": "de Guzman", "email": "rowel.deguzman+1@gmail.com", "userStatus": "1"}`)
+	var jsonStr = []byte(`{"firstName":"Rowel", "lastName": "de Guzman", "email": "` + faker.Email() + `", "password": "password1",  "userStatus": "1"}`)
 	request, _ := http.NewRequest("POST", "/v1/users/add", bytes.NewBuffer(jsonStr))
 	request.Header.Set("Content-Type", "application/json")
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		users.Create(w, r)
+		UserController.Create(w, r)
 	})
 
 	response := executeRequest(request, handler)
 
-	var m map[string]interface{}
+	var m map[string]any
 	err := json.Unmarshal(response.Body.Bytes(), &m)
 
 	if err != nil {
@@ -61,27 +82,45 @@ func TestCreates(t *testing.T) {
 		t.Errorf("Wrong status code. expecting 200, receive %v", m["statusCode"])
 	}
 
-	id, _ := m["devMessage"].(float64)
-	tearDown(strconv.Itoa(int(id)))
+	devMessage := m["devMessage"].(map[string]any)
+
+	fmt.Println(devMessage)
+	id, _ := devMessage["id"].(float64)
+	tearDown(t, strconv.Itoa(int(id)))
 
 }
 
-func TestGetDevice(t *testing.T) {
+func TestGetUsers(t *testing.T) {
 	request, _ := http.NewRequest("GET", "/v1/users/get", nil)
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		users.Get(w, r)
-	})
+	query := request.URL.Query()
+	query.Add("rows", "1")
+	query.Add("page", "2")
+	request.URL.RawQuery = query.Encode()
+
+	handler := http.HandlerFunc(UserController.Get)
 	response := executeRequest(request, handler)
 
-	var m map[string]interface{}
+	var m map[string]any
 	err := json.Unmarshal(response.Body.Bytes(), &m)
 
 	if err != nil {
 		t.Errorf("Unexpected response body found")
 	}
+
+	devMessage := m["devMessage"].([]any)
+	paginate := m["paginate"].(map[string]any)
+
 	if m["statusCode"] != float64(200) {
-		t.Errorf("Wrong status code. expecting 200, receive %v", m["statusCode"])
+		t.Errorf("Wrong status code. expecting 200, got %v", m["statusCode"])
 	}
 
-	log.Println(m)
+	if len(devMessage) > 1 {
+		t.Errorf("wrong query result. expecting %v. got %v", 1, len(devMessage))
+	}
+	if paginate["currentPage"] != float64(2) {
+		t.Errorf("wrong current page. expecting %v. got %v", 2, paginate["currentPage"])
+	}
+
+	assert.Equal(t, float64(2), paginate["currentPage"])
+
 }
